Rage Room stylization change
Capitilizing "rage room" leads the reader to believe that the phrase is said loudly, which is untrue. Further, I refuse to believe that the vet in this story came up with the phrase rage room. He purposefully describes that they are frightened when he displays The Pipe Strip so we should assume that they are canonically sane.